+++ lib/coinbase/transfer.rb
@@ -125,20 +125,19 @@ module Coinbase
     # Returns the link to the transaction on the blockchain explorer.
     # @return [String] The link to the transaction on the blockchain explorer
     def transaction_link
-      # TODO: Parameterize this by Network.
-      "https://sepolia.basescan.org/tx/#{transaction_hash}"
+      transaction.transaction_link
     end
 
     # Returns the Unsigned Payload of the Transfer.
     # @return [String] The Unsigned Payload
     def unsigned_payload
-      @model.unsigned_payload
+      transaction.unsigned_payload
     end
 
     # Returns the Signed Payload of the Transfer.
     # @return [String] The Signed Payload
     def signed_payload
-      @model.signed_payload
+      transaction.signed_payload
     end
 
     # Returns the Transfer transaction.
@@ -150,7 +149,7 @@ module Coinbase
     # Returns the Transaction Hash of the Transfer.
     # @return [String] The Transaction Hash
     def transaction_hash
-      @model.transaction_hash
+      transaction.transaction_hash
     end
 
     # Returns the status of the Transfer.